+++ src/adamicAdar.ts
